python/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
attrs==23.2.0
beautifulsoup4==4.13.4
blinker==1.8.2
cachetools==5.5.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
eloquent==0.5
exceptiongroup==1.2.1
faiss-cpu==1.11.0
filelock==3.18.0
Flask==3.0.3
flask-cors==6.0.1
fsspec==2025.5.1
google==3.0.0
google-auth==2.40.3
google-genai==1.21.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.7
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.5.1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
outcome==1.3.0.post0
packaging==24.0
pillow==11.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyMySQL==1.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
selenium==4.21.0
sentence-transformers==4.1.0
setuptools==78.1.1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
trio==0.25.1
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20240316
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.2.1
webdriver-manager==4.0.1
websockets==15.0.1
Werkzeug==3.0.4
wheel==0.45.1
wsproto==1.2.0
